problem_1/requirements.txt
together
python-dotenv
requests
